requirements.txt
# ============================================================
# Unsupervised Panoptic Pseudo-Label Generation — Requirements
# ============================================================
# Install: pip install -r requirements.txt
# NOTE: torch and torchvision are installed separately
#       by setup.sh (with correct CUDA/MPS index URLs).
# ============================================================

# ─── Vision & ML ───
scikit-learn>=1.3.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=9.2.0

# ─── DINOv2 / CAUSE feature extraction ───
timm==0.9.12
transformers>=4.30.0

# ─── CRF post-processing ───
pydensecrf

# ─── Cityscapes tools ───
cityscapesscripts

# ─── Misc ───
numpy>=1.21.0
tqdm
matplotlib

# ─── Stage-2: CUPS training (Detectron2 + PyTorch Lightning) ───
# Install detectron2 separately: python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'
pytorch-lightning>=2.0.0
kornia>=0.6.11
yacs
wandb>=0.15.4
rtpt==0.0.4
pycocotools
einops
torchmetrics==1.3.1
